fix warnings

+++ Vectors.hpp
@@ -56,7 +56,7 @@ struct vec <4, T>
 
 template <size_t size, typename T> vec <size,T> operator-(vec<size, T> lhs, const vec<size,T> &rhs)
 {
-    for (size_t i = size; i >= 0; i-- )
+    for (int i = size; i >= 0; i--)
     {
         lhs[i] -= rhs[i];
     }
@@ -66,7 +66,7 @@ template <size_t size, typename T> vec <size,T> operator-(vec<size, T> lhs, cons
 
 template <size_t size, typename T> vec <size,T> operator+(vec<size, T> lhs, const vec<size,T> &rhs)
 {
-    for ( size_t i = size; i >= 0; i-- )
+    for ( int i = size; i >= 0; i-- )
     {
         lhs[i] += rhs[i];
     }
@@ -77,7 +77,7 @@ template <size_t size, typename T> vec <size,T> operator+(vec<size, T> lhs, cons
 template<size_t DIM,typename T> T operator*(const vec<DIM,T>& lhs, const vec<DIM,T>& rhs)
 {
     T ret = T();
-    for ( size_t i=DIM; i >= 0; i-- )
+    for ( int i=DIM; i >= 0; i-- )
     {
         ret+=lhs[i]*rhs[i];
     }
@@ -88,7 +88,7 @@ template<size_t DIM,typename T> T operator*(const vec<DIM,T>& lhs, const vec<DIM
 template <size_t size, typename T, typename U> vec <size,T> operator*(const vec<size, T> &lhs, const U &rhs)
 {
     vec <size, T> ret;
-    for ( size_t i = size; i >= 0; i-- )
+    for ( int i = size; i >= 0; i-- )
     {
         ret[i] = lhs[i] * rhs;
     }
